requirements.txt
django==2.1.7
djangorestframework==3.9.2
gunicorn==19.9.0
psycopg2==2.7.7
dj-database-url==0.5.0


facebook-sdk==3.1.0
google-auth==1.6.3
google-api-python-client==1.7.8

django-cors-headers==2.5.2

ipython==7.5.0
